refactor: extract overlay helpers and remove unused menu dispatch
Make `dim_color` pub(crate) so it can be reused in the shell overlay.
Extract `dim_backdrop` and `render_text` helper functions from inline
rendering code in `shell_overlay.rs` to reduce duplication and improve
readability.

Remove the `dispatch_menu_action` function and any remaining unused
menu-handling code from the runtime. Menu actions are now dispatched
through a different path, so this dead code can safely be removed,
simplifying the runtime state machine.
