+++ lattice/src/compositor.rs
@@ -45,7 +45,7 @@ static DIM_TABLE: [u32; 256] = {
 
 /// Apply dim (~40% luminance) to a colour using the pre‑computed table.
 #[inline]
-fn dim_color(color: u32) -> u32 {
+pub(crate) fn dim_color(color: u32) -> u32 {
     let r = DIM_TABLE[((color >> 16) & 0xFF) as usize];
     let g = DIM_TABLE[((color >> 8) & 0xFF) as usize];
     let b = DIM_TABLE[(color & 0xFF) as usize];
